Re #34: IMonitoringProducer now managed per-server process
-PassthroughServer was the incorrect place for this as it means it existed per-stripe, not per-server
